Move preprocessor into extra project, Add description in examples
