--- core/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-com.linecorp.armeria.server.annotation.processor.DocumentationProcessor
